grabpublic/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.0.3
django-bootstrap3==14.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-notifications-hq==1.6.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
jsonfield==3.1.0
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.3
zipp==3.4.0
